requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0

# VQGAN dependencies
einops>=0.6.0
omegaconf>=2.3.0

# Gradio demo
gradio>=4.0.0

# Visualization
matplotlib>=3.7.0

# Training utilities
tqdm>=4.65.0

# Evaluation metrics
scikit-image>=0.20.0
opencv-python-headless>=4.7.0

# Config management
PyYAML>=6.0

# HuggingFace (checkpoint download)
huggingface_hub>=0.19.0
